Se corrigio el error por el cual no se incluia el archivo conversionFunctions.c y se debuguearon las funciones de conversion (salvo las de 64 bits) - 18/05/2023

+++ _system/inc/conversionFunctions.h
@@ -3,13 +3,13 @@ inline float ato_float(char* number);                           //string to floa
 inline double ato_double(char* number);                         //string to double
 
 inline void uint8_t_to_ascii(char* string, uint8_t number);     //uint8_t to string
-inline void uint16_t_to_ascii(char* string, uint8_t number);    //uint16_t to string
-inline void uint32_t_to_ascii(char* string, uint8_t number);    //uint32_t to string
-inline void uint64_t_to_ascii(char* string, uint8_t number);    //uint64_t to string
-inline void int8_t_to_ascii(char* string, uint8_t number);      //int8_t to string
-inline void int16_t_to_ascii(char* string, uint8_t number);     //int16_t to string
-inline void int32_t_to_ascii(char* string, uint8_t number);     //int32_t to string
-inline void int64_t_to_ascii(char* string, uint8_t number);     //int64_t to string
+inline void uint16_t_to_ascii(char* string, uint16_t number);   //uint16_t to string
+inline void uint32_t_to_ascii(char* string, uint32_t number);   //uint32_t to string
+inline void uint64_t_to_ascii(char* string, uint64_t number);   //uint64_t to string
+inline void int8_t_to_ascii(char* string, int8_t number);       //int8_t to string
+inline void int16_t_to_ascii(char* string, int16_t number);     //int16_t to string
+inline void int32_t_to_ascii(char* string, int32_t number);     //int32_t to string
+inline void int64_t_to_ascii(char* string, int64_t number);     //int64_t to string
 
 inline uint8_t ato_uint8(char* string);                         //string to uint8_t
 inline uint16_t ato_uint16(char* string);                       //string to uint16_t
